requirements.txt
mcp
pymysql
python-dotenv
psycopg2-binary